consolidate into single threaded publisher
