+++ bin/outboxer_publisher
@@ -8,22 +8,22 @@ require_relative '../app/jobs/event_created_job'
 
 options = {
   environment: ENV.fetch('OUTBOXER_ENV', 'development'),
-  buffer_size: ENV.fetch('OUTBOXER_BUFFER_SIZE', 200).to_i,
-  buffer_poll_interval: ENV.fetch('OUTBOXER_BUFFER_POLL_INTERVAL', 0.1),
-  database_poll_interval: ENV.fetch('OUTBOXER_DATABASE_POLL_INTERVAL', 10),
-  concurrency: ENV.fetch('OUTBOXER_CONCURRENCY', 1).to_i,
+  batch_size: ENV.fetch('OUTBOXER_BATCH_SIZE', 100),
+  poll_interval: ENV.fetch('OUTBOXER_POLL_INTERVAL', 5),
+  tick_interval: ENV.fetch('OUTBOXER_TICK_INTERVAL', 1),
   log_level: ENV.fetch('OUTBOXER_LOG_LEVEL', 'info').downcase,
-  redis_url: ENV.fetch('OUTBOXER_REDIS_URL', 'redis://localhost:6379/0'),
+  sidekiq_redis_url: ENV.fetch('SIDEKIQ_REDIS_URL', 'redis://localhost:6379/0'),
 }
 
 Sidekiq.configure_client do |config|
-  config.redis = { url: options[:redis_url], size: options[:concurrency] }
+  config.redis = { url: options[:sidekiq_redis_url], size: 1 }
 end
 
 logger = Outboxer::Logger.new $stdout, level: Logger.const_get(options[:log_level].upcase)
+logger.info "Outboxer v#{Outboxer::VERSION} starting in ruby #{RUBY_VERSION} " \
+            "(#{RUBY_RELEASE_DATE} revision #{RUBY_REVISION[0, 10]}) [#{RUBY_PLATFORM}]"
 
-db_config = Outboxer::Database.config(
-  environment: options[:environment], pool: options[:concurrency] + 1)
+db_config = Outboxer::Database.config(environment: options[:environment], pool: 1)
 
 Outboxer::Database.connect(config: db_config, logger: logger)
 
@@ -34,10 +34,8 @@ Outboxer::Database.connect(config: db_config, logger: logger)
 end
 
 Outboxer::Publisher.publish(
-  buffer_size: options[:buffer_size],
-  buffer_poll_interval: options[:buffer_poll_interval],
-  database_poll_interval: options[:database_poll_interval],
-  concurrency: options[:concurrency],
+  batch_size: options[:batch_size],
+  poll_interval: options[:poll_interval],
   logger: logger
 ) do |message|
   case message[:messageable_type]
