--- lib/outboxer/single_threaded_publisher.rb
@@ -1,93 +1,0 @@
-module Outboxer
-  module SingleThreadedPublisher
-    extend self
-
-    @publishing = false
-
-    def stop
-      @publishing = false
-    end
-
-    def sleep(duration, start_time:, tick_interval:, time:, kernel:)
-      while @publishing && (time.now - start_time) < duration
-        kernel.sleep(tick_interval)
-      end
-    end
-
-    def publish(
-      batch_size: 200, poll_interval: 5, tick_interval: 1,
-      logger: Logger.new($stdout, level: Logger::INFO),
-      time: Time, kernel: Kernel,
-      &block
-    )
-      logger.info "Dequeueing up to #{batch_size} messages every #{poll_interval} second(s)"
-
-      @publishing = true
-
-      while @publishing
-        begin
-          dequeued_messages = Messages.dequeue(limit: batch_size)
-
-          dequeued_messages.each do |dequeued_message|
-            publish_message(
-              dequeued_message: { id: dequeued_message[:id], dequeued_at: time.now.utc },
-              logger: logger, time: time, kernel: kernel,
-              &block)
-          end
-
-          if dequeued_messages.count < batch_size
-            Publisher.sleep(
-              poll_interval, start_time: time.now, tick_interval: tick_interval,
-              time: time, kernel: kernel)
-          end
-        rescue StandardError => exception
-          logger.error "#{exception.class}: #{exception.message}"
-          exception.backtrace.each { |frame| logger.error frame }
-        rescue Exception => exception
-          logger.fatal "#{exception.class}: #{exception.message}"
-          exception.backtrace.each { |frame| logger.fatal frame }
-
-          stop
-        end
-      end
-
-      logger.info "Stopped dequeueing messages"
-    end
-
-    def publish_message(dequeued_message:, logger:, time:, kernel:, &block)
-      dequeued_at = dequeued_message[:dequeued_at]
-
-      message = Message.publishing(id: dequeued_message[:id])
-      logger.debug "Publishing message #{message[:id]} for " \
-        "#{message[:messageable_type]}::#{message[:messageable_id]} " \
-        "in #{(time.now.utc - dequeued_at).round(3)}s"
-
-      begin
-        block.call(message)
-      rescue Exception => exception
-        Message.failed(id: message[:id], exception: exception)
-        logger.error "Failed to publish message #{message[:id]} for " \
-          "#{message[:messageable_type]}::#{message[:messageable_id]} " \
-          "in #{(time.now.utc - dequeued_at).round(3)}s"
-
-        raise
-      end
-
-      Message.published(id: message[:id])
-      logger.debug "Published message #{message[:id]} for " \
-        "#{message[:messageable_type]}::#{message[:messageable_id]} " \
-        "in #{(time.now.utc - dequeued_at).round(3)}s"
-    rescue StandardError => exception
-      logger.error "#{exception.class}: #{exception.message} " \
-        "in #{(time.now.utc - dequeued_at).round(3)}s"
-
-      exception.backtrace.each { |frame| logger.error frame }
-    rescue Exception => exception
-      logger.fatal "#{exception.class}: #{exception.message} " \
-        "in #{(time.now.utc - dequeued_at).round(3)}s"
-      exception.backtrace.each { |frame| logger.error frame }
-
-      stop
-    end
-  end
-end
