+++ spec/lib/outboxer/publisher/publish_spec.rb
@@ -3,9 +3,8 @@ require 'spec_helper'
 module Outboxer
   RSpec.describe Publisher do
     describe '.publish' do
-      let(:buffer_size) { 1 }
+      let(:batch_size) { 1 }
       let(:poll_interval) { 1 }
-      let(:concurrency) { 1 }
       let(:logger) { instance_double(Logger, debug: true, error: true, fatal: true, info: true) }
       let(:kernel) { class_double(Kernel, sleep: nil) }
 
@@ -16,11 +15,10 @@ module Outboxer
       let!(:queued_message) { create(:outboxer_message, :queued) }
 
       context 'when message published successfully' do
-        it 'deletes existing message' do
+        it 'sets the message to published' do
           Publisher.publish(
-            buffer_size: buffer_size,
+            batch_size: batch_size,
             poll_interval: poll_interval,
-            concurrency: concurrency,
             logger: logger,
             kernel: kernel
           ) do |message|
@@ -42,9 +40,8 @@ module Outboxer
 
           before do
             Publisher.publish(
-              buffer_size: buffer_size,
+              batch_size: batch_size,
               poll_interval: poll_interval,
-              concurrency: concurrency,
               logger: logger,
               kernel: kernel
             ) do |message|
@@ -58,13 +55,11 @@ module Outboxer
             queued_message.reload
 
             expect(queued_message.status).to eq(Models::Message::Status::FAILED)
-
             expect(queued_message.exceptions.count).to eq(1)
             expect(queued_message.exceptions[0].class_name).to eq(standard_error.class.name)
             expect(queued_message.exceptions[0].message_text).to eq(standard_error.message)
             expect(queued_message.exceptions[0].created_at).not_to be_nil
 
-            expect(queued_message.exceptions[0].frames.count).to eq(4)
             expect(queued_message.exceptions[0].frames[0].index).to eq(0)
 
             expect(queued_message.exceptions[0].frames[0].text).to match(
@@ -88,13 +83,12 @@ module Outboxer
 
           before do
             Publisher.publish(
-              buffer_size: buffer_size,
+              batch_size: batch_size,
               poll_interval: poll_interval,
-              concurrency: concurrency,
               logger: logger,
               kernel: kernel
             ) do |dequeued_message|
-              Publisher.stop
+              Publish.stop
 
               raise no_memory_error
             end
@@ -104,13 +98,11 @@ module Outboxer
             queued_message.reload
 
             expect(queued_message.status).to eq(Models::Message::Status::FAILED)
-
             expect(queued_message.exceptions.count).to eq(1)
             expect(queued_message.exceptions[0].class_name).to eq(no_memory_error.class.name)
-            expect(queued_message.exceptions[0].message_text).to eq(no_memory_error.message)
+            expect(queued_message.exceptions[0].message_text). to eq(no_memory_error.message)
             expect(queued_message.exceptions[0].created_at).not_to be_nil
 
-            expect(queued_message.exceptions[0].frames.count).to eq(4)
             expect(queued_message.exceptions[0].frames[0].index).to eq(0)
             expect(queued_message.exceptions[0].frames[0].text).to match(
               /outboxer\/publisher\/publish_spec.rb:\d+:in `block \(6 levels\) in <module:Outboxer>'/)
@@ -139,7 +131,7 @@ module Outboxer
               when 1
                 raise StandardError, 'queue error'
               else
-                Publisher.stop
+                Publish.stop
 
                 []
               end
@@ -148,8 +140,7 @@ module Outboxer
             expect(logger).to receive(:error).with(include('StandardError: queue error')).once
 
             Publisher.publish(
-              buffer_size: buffer_size,
-              concurrency: concurrency,
+              batch_size: batch_size,
               poll_interval: poll_interval,
               logger: logger,
               kernel: kernel
@@ -167,11 +158,11 @@ module Outboxer
               .once
 
             Publisher.publish(
-              buffer_size: buffer_size,
+              batch_size: batch_size,
               poll_interval: poll_interval,
-              concurrency: concurrency,
               logger: logger,
-              kernel: kernel)
+              kernel: kernel
+            )
           end
         end
       end
